+++ test/src/extractor/test_looper_layer.cpp
@@ -14,7 +14,7 @@ struct dummy_extractor
 {
     stub::function<uint64_t()> decoding_timestamp;
     stub::function<uint64_t()> presentation_timestamp;
-    stub::function<uint64_t()> video_length;
+    stub::function<uint64_t()> media_duration;
     stub::function<void()> advance;
     stub::function<void()> reset;
     stub::function<bool()> at_end;
@@ -35,7 +35,7 @@ TEST(extractor_test_looper_layer, not_looping)
     extractor.advance();
     extractor.advance();
     EXPECT_EQ(0U, layer.reset.calls());
-    EXPECT_EQ(0U, layer.video_length.calls());
+    EXPECT_EQ(0U, layer.media_duration.calls());
     EXPECT_EQ(1U, extractor.decoding_timestamp());
     EXPECT_EQ(1U, extractor.presentation_timestamp());
     EXPECT_EQ(3U, layer.advance.calls());
@@ -47,14 +47,14 @@ TEST(extractor_test_looper_layer, looping)
     dummy_extractor& layer = extractor;
     layer.decoding_timestamp.set_return(1);
     layer.presentation_timestamp.set_return(1);
-    layer.video_length.set_return(2);
+    layer.media_duration.set_return(2);
     layer.at_end.set_return(false, false, true, false, false, true);
     EXPECT_EQ(0U, extractor.loops());
     extractor.advance();
     extractor.advance();
     extractor.advance();
     EXPECT_EQ(1U, layer.reset.calls());
-    EXPECT_EQ(1U, layer.video_length.calls());
+    EXPECT_EQ(1U, layer.media_duration.calls());
     EXPECT_EQ(3U, extractor.decoding_timestamp());
     EXPECT_EQ(3U, extractor.presentation_timestamp());
     EXPECT_EQ(3U, layer.advance.calls());
@@ -63,7 +63,7 @@ TEST(extractor_test_looper_layer, looping)
     extractor.advance();
     extractor.advance();
     EXPECT_EQ(2U, layer.reset.calls());
-    EXPECT_EQ(2U, layer.video_length.calls());
+    EXPECT_EQ(2U, layer.media_duration.calls());
     EXPECT_EQ(5U, extractor.decoding_timestamp());
     EXPECT_EQ(5U, extractor.presentation_timestamp());
     EXPECT_EQ(6U, layer.advance.calls());